CASdatasets package

Actuarial Datasets (originally for the 'Computational Actuarial Science with R' book)

Details

This package contains aggregated and policy-level datasets. Below a list by country or region is given.

Australia:

auscathist: Historical disaster statistics in Australia.

ausNLHYCapAdeq, ausNLHYFinPerf, ausNLHYFinPos, ausNLHYLiability, ausNLHYOffProf, ausNLHYOpIncExp, ausNLHYPremClaim, ausNLHYPrivInsur, ausNLHYPubInsur, ausNLHYRecAASB, ausNLHYReserve: Australian Market - non-life insurance (aggregate level).

ausNLHYCapAdeqByComp, ausNLHYClaimByState, ausNLHYFinPerfByComp, ausNLHYFinPerfPublic, ausNLHYFinPosByComp, ausNLHYFinPosPublic, ausNLHYOpIncExpPublic, ausNLHYPremByState, ausNLHYPremClaimPublic, ausNLHYPrivInsur, ausNLHYPubInsur: Australian Market - non-life insurance (company, state, public level).

ausNLHYLloydAsset, ausNLHYLloydGPI ausNLHYLloydUWAcc, ausNLHYLloydUWRes: Lloyds Market in Australia.

austri1autoBI7895, austri2auto: Australian claim triangles.

ausprivautolong, ausMTPL8486, ausprivauto0405: Australian private motor datasets (by policy).

ausNSWdriver04, ausNSWdeath02: New South Wales statistics.

Belgium:

beaonre: AON Re Belgium dataset.

beMTPL16: a MTPL Belgium dataset of 2016 (by policy).

beMTPL97: a MTPL Belgium dataset of 1997 (by policy).

besecura: Secura Re Belgium dataset.

Brazil:

brgeomunicins: a dataset of Brazilian cities with insurance indicators (exposure, claim frequency and premium). Geographical datasets brgeomunic_sf, brgeomunic_sp should be downloaded from the github repository in extra directory geodata.

brautocoll: a Brazilian dataset of auto collision in Belo Horizonte during one month (categorical claim severity).

braggprem, braggclaim: aggregate Brazilian datasets per region.

brvehins1: a Brazilian vehicle insurance datasets (by policy) with risk features except City code.

brvehins2: a Brazilian vehicle insurance datasets (by policy) with risk features including City code.

Canada:

canlifins, canlifins2: A portfolio of a Canadian life insurer.

catelematic13: Canadian telematic insurance dataset.

Danemark:

danishuni, danishmulti: Danish reinsurance claim dataset.

European Union:

eudirectlapse: non-life insurance lapse dataset.

euhealthinsurance: European health insurance dataset.

euMTPL: European motor TPL insurance dataset.

France:

freaggnumber: a French aggregate claim number dataset.

frebiloss: French business interruption losses.

freclaimset, freclaimset9207: French claim settlements.

frecomfire: French commercial fire losses.

freDisTables: French disability tables.

fremarine: French marine claim dataset (by policy).

freMortTables: French mortality tables.

fremotorclaim: French private motor claim datasets (by policy).

freMTPL: two French Motor-TPL claim datasets (by policy).

freMPL: a collection of ten French Motor personal line datasets (by policy).

freportfolio: fictive mortality tables and French nation-wide corresponding tables; two disability datasets from a French insurer.

fretelematic: a French telematic Motor-TPL claim dataset (by policy).

fretplclaimnumber: a French Motor-TPL claim dataset (by policy).

fre4LoBtriangles: A collection of triangles line of business from a private insurer.

pricingame: Datesets used at Pricing Games of the French Institute of Actuaries.

Germany:

credit: A German Credit dataset.

Italy:

itamtplcost: Large losses of an Italian Motor-TPL company.

New Zealand:

nzcathist: Historical disaster statistics in New Zealand.

Norway:

norauto: Norwegian automobile dataset.

norfire: Norwegian fire dataset.

Norberg: Norberg's credibility dataset.

nortritpl8800: Norwegian claim triangle.

Singapore:

sgautonb: Singapore Automobile claim count dataset.

sgtriangles: Singapore Property and Casualty triangles.

Sweden:

swautoins: Swedish Motor Insurance dataset

swbusscase: Swedish Buss Insurance dataset

swmotorcycle: Swedish Motorcycle Insurance dataset

United Kingdom:

ukaggclaim: United Kingdom Car Insurance Claims.

ukautocoll: United Kingdom Car Collision Insurance Claims.

United States of America:

Davis: Davis height-weight dataset.

ICB1, ICB2: Insurance Company Benchmarks.

lossalae,lossalaefull: General Third Part-liability claims and expenses.

SOAGMI: SOA Group Medical Insurance dataset.

usautoBI: Automobile Bodily Injuries in US.

usautotriangles: US automobile triangles.

usexpense: US expense dataset.

usGLtriangles: US Property and Casualty triangles.

ushurricane, ushustormloss4980: Historical hurricane statistics in United States of America.

uslapseagent: US lapse dataset from tied-agent channel.

usmassBI: US Massachusetts Automobile bodily injury claim datasets.

usmedclaim: US medical claim triangle.

usMSHA1316: US Mine Safety and Health Administration claim dataset.

usMVTA: US motor vehicle traffic accident.

usprivautoclaim: private automobile claims.

usquakeLR: California earthquake loss ratios.

ustermlife: Term life insurance survey.

uswarrantaggnum: US warranty automobile.

usworkcomp: US workers compensation datasets.

Misc.:

eqlist: Earthquake list.

hurricanehist: Hurricane history.

PnCdemand: Property and Casualty demand.

spacedata: Space dataset.

ECBYieldCurve, FedYieldCurve Yield curve for eurozone and US.

forexUSUK Foreign exchange rate between USD and GBP.

Use the HMD website https://www.mortality.org/ for mortality tables.

Here is a list of datasets whose name has changed compared to the book 'Computational Actuarial Science with R':'

Chapter 1: extreme2datasince1899 is hurricanehist.

Chapter 5: accidents and accidents_data are merged in brautocoll; 55mu2500gsd is brgeomunic_sf in extra directory geodata in the github repository; sul_sp is stored in brgeomunicins.

Chapter 9: MyPortfolio is freprojqxINSEE.

Chapter 10: DataMortality is freptfpermdis.

Chapter 11: DEXUSUK is forexUSUK.

Chapter 14: CONTRACTS is freMTPLfreq; CLAIMS is freMTPLsev.

Chapter 15: AutoClaimData is usmassBI2.

Author

Arthur Charpentier, Christophe Dutang.