Australian Market - non-life insurance (Glossary)

Financial performance and financial position of insurers operating in Australia between 2005 and 2010 (Glossary).

Details

Glossary notes:

Capital base is the amount of eligible capital held by an insurer to provide a buffer against losses that have not been anticipated and, in the event of problems, enable the insurer to continue operating while those problems are addressed or resolved. For locally incorporated insurers it is the sum of tier 1 capital (net of deductions) and tier 2 capital . Capital base for branch insurers is derived from net assets inside Australia.

Captive insurer is a company within a group of related companies performing the function of insurer to that group.

Classes of business in tables 7-11 are shown in order of risk capital factors as described in guidance note GGN 110.3.

Direct insurers are those insurers who, excluding intra-group arrangements, predominantly undertake liability by way of direct insurance business.

Earned premium (as defined in AASB 1023 ) is the amount of premium earned during the financial year and includes movements in the unearned premium provision.

Gross claims expense (as per table 11) relates to: claims that are paid during a financial period; and recognised claims liabilities (i.e. movement in outstanding claims provision).

Gross incurred claims comprises claims paid during the period, movements in the outstanding claims provision and movements in premium liabilities .

Gross premium revenue is recognised fully when the business is written. The accounting concepts of earned and unearned premium are no longer recognised under the APRA prudential framework, hence this item is not consistent with AASB 1023 requirements. Instead, the potential claims liabilities arising from the uncovered term of written insurance business are recognised through the creation of premium liabilities .

LMI (Lenders mortgage insurers) provide cover to protect lenders from default by borrowers on loans secured by mortgage. Mortgage insurers are substantially different to other insurers and are subject to special condition of authority.

Lower tier 2 ratio is lower tier 2 capital divided by tier 1 capital (net of deductions) . The regulatory maximum for this ratio is 50 percent.

Lloyd's is a London based insurance market in which business is underwritten by both individuals and corporate members who form syndicates to accept risk.

Minimum capital requirement is the amount of risk-based capital APRA requires general insurers to hold to meet its insurance obligations under a wide range of circumstances.

Net incurred claims is gross incurred claims net of reinsurance recoveries revenue and non-reinsurance recoveries revenue.

Net loss ratio is net incurred claims divided by net premium revenue. Net premium revenue is gross premium revenue net of outwards reinsurance expense.

Net profit/loss refers to profit or loss from ordinary activities after income tax, before extraordinary items.

Non-reinsurance recoverables comprise recoverables from subrogation, salvage, sharing arrangements etc, net of provision for doubtful debts.

Non-reinsurance recoveries revenue comprises amounts the insurer has recovered or is entitled to recover from subrogation, salvage and other non-reinsurance recoveries.

Other assets comprises investment income receivable, other reinsurance assets receivable from reinsurers (i.e. other than reinsurance recoveries), GST receivable, other receivables, tax assets, plant and equipment (net of depreciation) and other assets.

Other investments are strategic investments/acquisitions and other investments that do not constitute investments integral to insurance operations.

Other items comprises other operating income, goodwill amortisation and income tax expense or benefit. Other liabilities comprises creditors and accruals, other provisions and other liabilities. Other operating expenses are all operating expenses not related to underwriting.

Outstanding claims provision is the insurer's liability for outstanding claims. It recognises the potential cost to the insurer of settling claims which it has incurred at the reporting date (including estimates of claims that have not yet been notified to the insurer), but which have not been paid. The amount reported is after taking account of inflation and discounting, without deducting reinsurance and non- reinsurance recoverables .

Outwards reinsurance expense is premium ceded to reinsurers, recognised as an expense fully when incurred or contracted.

Payables on reinsurance contracts comprise amounts payable to reinsurers. This includes premiums payable but not yet due for payment, deposits withheld from reinsurers, commissions due to reinsurers and the reinsurers' portion of recoveries and salvage.

Premium liabilities relate to the future claims arising from future events insured under existing policies accepted. This fully prospective determination is a more effective means of recognising potential risk than the accounting concept of unearned premium. The amount reported is after taking `account of inflation and discounting, without deducting reinsurance and non-reinsurance recoveries.

Premium receivables are premiums due, net of provision for doubtful debts, including unclosed business written close to the reporting date.

Reinsurance recoverables comprise amounts recoverable under reinsurance contracts. Reinsurance and other recoverables is the aggregate of reinsurance recoverables and non-reinsurance recoverables.

Reinsurance recoveries revenue comprises amounts the insurer has recovered or is entitled to recover from reinsurers on incurred claims during the reporting period.

Reinsurers are those insurers who, excluding intra-group arrangements, predominantly undertake liability by way of reinsurance business.

Return on assets is net profit/loss divided by the average on-balance sheet total assets for the period. Return on equity is net profit/loss divided by the average shareholders' equity for the period.

Run-off insurers are restricted by APRA from writing new or renewal insurance business. However, the company may still be acting as an insurance agent, broker or underwriting agent for other general insurers.

Solvency coverage is capital base divided by minimum capital requirement.

Tier 1 capital (net of deductions) comprises the highest quality capital elements, including: paid-up ordinary shares, general reserves, retained earnings, current year earnings net of expected dividends and tax expenses, technical provisions in excess of those required by GPS 210 , non-cumulative irredeemable preference shares and other "innovative" capital instruments. This amount is net of goodwill, other intangible assets and future income tax benefits.

Source

Data is copyrighted by Australian Prudential Regulation Authority (APRA) and is under the Creative Commons - By licence. Please refer to https://www.apra.gov.au/

See also