European motor TPL insurance

The euMTPL compiles three years of experience from a European MTPL (Motor Third Party Liability) portfolio, including frequency and severity values for different types of losses. The data was collected during the first decade of the 21st century.

Usage

data(euMTPL)

Format

euMTPL is a data frame with 2,373,197 rows and 19 columns:

policy_id

Unique identifier for each policy.

year

Calendar year of the policy.

group

Data split into training, validation, and test sets using a 70/10/20 ratio.

fuel_type

Fuel type of the insured vehicle.

vehicle_category

Category of the insured vehicle.

vehicle_use

Intended use of the vehicle (e.g., personal, commercial).

province

Province of residence of the policyholder.

horsepower

Power output of the insured vehicle, measured in horsepower.

gender

Gender of the policyholder.

age

Age of the policyholder at the start date of the policy.

exposure

Fraction of the year that the policy was in effect.

cost_nc

Total claim amount for No Card (NC) claims.

num_nc

Number of No Card (NC) claims.

cost_cg

Total claim amount for Card Gestionario (CG) claims.

num_cg

Number of Card Gestionario (CG) claims.

cost_cd

Total claim amount for Card Debitore (CD) claims.

num_cd

Number of Card Debitore (CD) claims.

cost_fcd

Total claim amount for Forfait Card Gestionario (FCD) claims.

num_fcd

Number of Forfait Card Gestionario (FCD) claims.

Source

Unknown non-life insurers from European Union.

Examples

# (1) load of data
#
data(euMTPL)
head(euMTPL)
#> # A tibble: 6 × 19
#>   policy_id group fuel_type  year vehicle_category vehicle_use province
#>       <int> <chr> <fct>     <dbl> <fct>            <fct>       <fct>   
#> 1         1 test  B             7 1                1           PA      
#> 2         2 train B             7 1                1           NA      
#> 3         4 train B             7 1                1           CN      
#> 4         5 train B             7 1                1           NA      
#> 5         6 train B             7 1                1           NA      
#> 6         8 train B             7 1                1           NA      
#> # ℹ 12 more variables: horsepower <int>, gender <fct>, age <int>,
#> #   exposure <dbl>, cost_nc <dbl>, num_nc <int>, cost_cg <dbl>, num_cg <int>,
#> #   cost_fcg <dbl>, num_fcg <int>, cost_cd <dbl>, num_cd <int>