European health insurance

The euhealthinsurance compiles data coming from a health group collective fund that covers different kind of health perils to the members. Available data are: gender, age at inception of coverage, role in the policy, number and agggregate amount.

Usage

data(euhealthinsurance)

Format

euhealthinsurance is a dataframe with 157221 observations and 21 columns

id_ano

anonymized id.

relation

role in the policy.

gender

gender: M, F.

policy_years

cumulated exposure.

age_at_inception

attained age when the policy started.

num_analysis

number of laboratory test covered.

num_dentistry

number of dental health services covered.

num_diagnostics

number of exams covered.

num_endoscopy

number of endoscopies covered.

num_hospitalizations

number of hospitalizations covered.

num_mammography

number of mammographies covered.

num_operations

number of surgeries covered.

num_visits

number of specialist visits covered.

amt_analysis

amount of laboratory test covered.

amt_dentistry

amount of dental health services covered.

amt_diagnostics

amount of exams covered.

amt_endoscopy

amount of endoscopies covered.

amt_hospitalizations

amount of hospitalizations covered.

amt_mammography

amount of mammographies covered.

amt_operations

amount of surgeries covered.

amt_visits

amount of specialist visits covered.

Source

Unknown non-life insurers from European Union.

Examples

# (1) load of data
#
data(euhealthinsurance)
head(euhealthinsurance)
#> # A tibble: 6 × 22
#>   id_anon relation       gender policy_years age_at_inception num_analysis
#>   <fct>   <fct>          <fct>         <dbl>            <dbl>        <int>
#> 1 CA0001  contract_owner M              1.50             46.1            0
#> 2 CA0002  contract_owner M              1.50             33.5            0
#> 3 CA0003  contract_owner M              1.50             41.4            0
#> 4 CA0004  contract_owner M              1.50             36.0            0
#> 5 CA0005  contract_owner M              1.50             54.2            0
#> 6 CA0006  contract_owner M              1.50             50.1            0
#> # ℹ 16 more variables: num_dentistry <int>, num_diagnostics <int>,
#> #   num_endoscopy <int>, num_hospitalizations <int>, num_mammography <int>,
#> #   num_operations <int>, num_visits <int>, amt_analysis <dbl>,
#> #   amt_dentistry <dbl>, amt_diagnostics <dbl>, amt_endoscopy <dbl>,
#> #   amt_hospitalizations <dbl>, amt_mammography <dbl>, amt_operations <dbl>,
#> #   amt_visits <dbl>, group <fct>