French geospatial dataset for departments

fredpt17 is a spatial database containing geospatial information of French departments to be used with pricing actuarial games' files, see pricingame. fredpt17 is a geospatial dataframe of class sp based on six files: DEPARTMENTS.cpg, DEPARTMENTS.dbf, DEPARTMENTS.prj, DEPARTMENTS.qpj, DEPARTMENTS.shp, DEPARTMENTS.shx. As it is of class sp, fredpt17 can be easily plotted or summarized.

Usage

data(fredpt17)

Format

fredpt17@data contains 1 column:

DEPT

A factor for the department number.

Source

See https://actinfo.hypotheses.org/86 for the third pricing game.

See also

Examples

# (1) load of data
#
data(fredpt17)

names(fredpt17)
#> [1] "DEPT"
class(fredpt17)
#> [1] "SpatialPolygonsDataFrame"
#> attr(,"package")
#> [1] "sp"
length(fredpt17)
#> [1] 96
summary(fredpt17)
#> Object of class SpatialPolygonsDataFrame
#> Coordinates:
#>         min     max
#> x   99217.1 1242375
#> y 6049680.0 7110480
#> Is projected: TRUE 
#> proj4string :
#> [+proj=lcc +lat_1=45.898918964419 +lat_2=47.696014502038 +lat_0=46.8
#> +lon_0=0 +x_0=600000 +y_0=2200000 +a=6378249.2 +b=6356515.000000472
#> +towgs84=-168,-60,320,-0,-0,-0,0 +pm=2.337229166667 +units=m +no_defs]
#> Data attributes:
#>       DEPT   
#>  01     : 1  
#>  02     : 1  
#>  03     : 1  
#>  04     : 1  
#>  05     : 1  
#>  06     : 1  
#>  (Other):90  
dim(fredpt17)
#> [1] 96  1

if (FALSE) {
# (2) plot of data
#
 
 plot(fredpt17)
 
 spplot(fredpt17)
 

}