Portfolio Mortality and Disability Tables

The freprojqxINSEE table has been established on INSEE projection for the period 2007-2060 based a median scenario, cf. Blanpain and Chardon (2010), adjusted and selected for the purpose of the book.

The frefictivetable represents a fictive portfolio of 87,090 individuals that enter in a healthy condition and have been observed between 1996-01-01 and 2007-12-31. The exit (that may occur before December 2007) is either "deceased" or "other".

The frefictivetable2,frefictivetable3 represents a fictive portfolio of 100,000 individuals that enter in a healthy condition and have been observed between December 1988 and December 1998. The exit is either "deceased" or "other" for censored observation.

The freptfpermdis and freptftempdis datasets comes from two portfolio of two French private companies (insurer or institute), respectively for permanent disability insurance and temporary disability insurance.

Usage

data(freprojqxINSEE)
data(frefictivetable)
data(frefictivetable2)
data(frefictivetable3)
data(freptfpermdis)
data(freptftempdis)

Format

freprojqxINSEE is a data frame of 109 columns and 66 rows:

Age

The age.

F2007,..., F2060

The 1-year female death probabilities

M2007,..., M2060

The 1-year male death probabilities

frefictivetable is a data frame of 6 columns and 87,090 rows:

Id

the identification number.

Gender

the gender as "factor".

DateOfBirth

the date of birth as "Date".

DateIn

the entry date as "Date".

DateOut

the exit date as "Date".

Status

the status at exit : "deceased" (i.e. non-censored observation) or "other" (i.e. censored observation) as "factor".

frefictivetable2,frefictivetable3 are data frames of 5 columns and 100,000 rows:

DateIn

the entry date as "Date".

DateOut

the exit date as "Date".

Status

the status at exit : "deceased" (i.e. non-censored observation) or "other" (i.e. censored observation) as "factor".

DateOfBirth

the date of birth as "Date".

Gender

the gender as "factor".

freptfpermdis is a data frame of 6 columns and 1,048,575 rows:

PolicyID

the policy identification number.

BirthDate

the date of birth.

Gender

the sex: M for male and F for female.

EntryDate

the entry date.

ExitDate

the exit date.

ExitStatus

the status at exit: "deceased" (i.e. non-censored observation) or "other" (i.e. censored observation).

freptftempdis is a data frame of 9 columns and 560,725 rows:

Gender

the sex: M for male and F for female.

JobType

the job category: "employee", "managers, engineers, sales responsibles", "non-manager employee" "other 1", "other 2", "other 3", "other 4", "other 5", "technician", "unemployed workers".

UWType

the underwriting type: either "specific policy in a collective agreement", "specific policy not linked to a collective agreement", "standard policy in a collective agreement" or "standard policy not linked to a collective agreement".

JobStopType

the reason for disability: "illness", "work accident", "pregnancy" (for women only).

Birthdate

the date of birth.

OccurDate

the date of occurence.

EntryDate

the entry date.

ExitDate

the exit date.

JobComebackType

the status at exit: "recovered" (i.e. non-censored observation: the person goes back to work), "disabled" (i.e. non-censored observation: the person is permanently disabled) or "on-going" (i.e. censored observation).

Source

For freprojqxINSEE, Blanpain and Chardon (2010).

For frefictivetable, Chapter 9 of Computational Actuarial Science with R, Ed. Arthur Charpentier, Chapman and Hall/CRC The R Series, 2014.

For freptfpermdis, freptftempdis, RessourcesActuarielles

References

Blanpain, N. and Chardon, O. (2010). Projections de populations 2007-2060 pour la France metropolitaine: methode et principaux resultats. Serie des Documents de Travail de la direction des statistiques Demographiques et Sociales F1008, INSEE.

Examples

# (1) load of data
#
data(freprojqxINSEE)
data(frefictivetable)

head(freprojqxINSEE)
#>   Age        F2007        F2008        F2009        F2010        F2011
#> 1  30 0.0003115935 0.0003060422 0.0003004796 0.0002949099 0.0002893372
#> 2  31 0.0003483465 0.0003421842 0.0003360418 0.0003299222 0.0003238283
#> 3  32 0.0003890368 0.0003822279 0.0003754724 0.0003687718 0.0003621274
#> 4  33 0.0004340360 0.0004265487 0.0004191498 0.0004118392 0.0004046167
#> 5  34 0.0004837451 0.0004755527 0.0004674842 0.0004595380 0.0004517125
#> 6  35 0.0005385961 0.0005296784 0.0005209197 0.0005123170 0.0005038673
#>          F2012        F2013        F2014        F2015        F2016        F2017
#> 1 0.0002837657 0.0002781994 0.0002726422 0.0002670980 0.0002615706 0.0002560635
#> 2 0.0003177627 0.0003117282 0.0003057273 0.0002997624 0.0002938361 0.0002879507
#> 3 0.0003555404 0.0003490121 0.0003425438 0.0003361364 0.0003297912 0.0003235091
#> 4 0.0003974822 0.0003904355 0.0003834764 0.0003766046 0.0003698200 0.0003631223
#> 5 0.0004440062 0.0004364176 0.0004289450 0.0004215871 0.0004143422 0.0004072089
#> 6 0.0004955677 0.0004874155 0.0004794079 0.0004715421 0.0004638154 0.0004562253
#>          F2018        F2019        F2020        F2021        F2022        F2023
#> 1 0.0002505805 0.0002451249 0.0002397001 0.0002343094 0.0002289559 0.0002236427
#> 2 0.0002821084 0.0002763114 0.0002705618 0.0002648617 0.0002592130 0.0002536176
#> 3 0.0003172912 0.0003111382 0.0003050512 0.0002990309 0.0002930782 0.0002871936
#> 4 0.0003565112 0.0003499865 0.0003435476 0.0003371945 0.0003309266 0.0003247436
#> 5 0.0004001857 0.0003932712 0.0003864640 0.0003797625 0.0003731653 0.0003666712
#> 6 0.0004487692 0.0004414446 0.0004342489 0.0004271799 0.0004202351 0.0004134123
#>          F2024        F2025        F2026        F2027        F2028        F2029
#> 1 0.0002183726 0.0002131484 0.0002079730 0.0002028487 0.0001977781 0.0001927635
#> 2 0.0002480772 0.0002425936 0.0002371684 0.0002318030 0.0002264990 0.0002212578
#> 3 0.0002813779 0.0002756317 0.0002699557 0.0002643502 0.0002588158 0.0002533529
#> 4 0.0003186451 0.0003126307 0.0003067000 0.0003008525 0.0002950878 0.0002894054
#> 5 0.0003602787 0.0003539863 0.0003477929 0.0003416970 0.0003356973 0.0003297925
#> 6 0.0004067090 0.0004001232 0.0003936526 0.0003872951 0.0003810486 0.0003749110
#>          F2030        F2031        F2032        F2033        F2034        F2035
#> 1 0.0001878072 0.0001829111 0.0001780773 0.0001733077 0.0001686181 0.0001640912
#> 2 0.0002160806 0.0002109686 0.0002059230 0.0002009448 0.0001960472 0.0001913009
#> 3 0.0002479619 0.0002426431 0.0002373969 0.0002322234 0.0002271325 0.0002221803
#> 4 0.0002838047 0.0002782854 0.0002728469 0.0002674886 0.0002622167 0.0002570696
#> 5 0.0003239812 0.0003182623 0.0003126345 0.0003070964 0.0003016502 0.0002963141
#> 6 0.0003688803 0.0003629545 0.0003571317 0.0003514100 0.0003457872 0.0003402597
#>          F2036        F2037        F2038        F2039        F2040        F2041
#> 1 0.0001597342 0.0001555399 0.0001515015 0.0001476126 0.0001438671 0.0001402590
#> 2 0.0001867124 0.0001822758 0.0001779854 0.0001738359 0.0001698221 0.0001659390
#> 3 0.0002173717 0.0002127021 0.0002081670 0.0002037620 0.0001994830 0.0001953259
#> 4 0.0002520503 0.0002471555 0.0002423815 0.0002377250 0.0002331829 0.0002287520
#> 5 0.0002910890 0.0002859723 0.0002809615 0.0002760542 0.0002712480 0.0002665406
#> 6 0.0003348256 0.0003294831 0.0003242306 0.0003190666 0.0003139895 0.0003089978
#>          F2042        F2043        F2044        F2045        F2046        F2047
#> 1 0.0001367828 0.0001334330 0.0001302047 0.0001270929 0.0001240930 0.0001212006
#> 2 0.0001621819 0.0001585462 0.0001550276 0.0001516219 0.0001483250 0.0001451331
#> 3 0.0001912868 0.0001873620 0.0001835478 0.0001798408 0.0001762376 0.0001727350
#> 4 0.0002244291 0.0002202115 0.0002160961 0.0002120803 0.0002081612 0.0002043364
#> 5 0.0002619298 0.0002574134 0.0002529893 0.0002486553 0.0002444095 0.0002402498
#> 6 0.0003040900 0.0002992645 0.0002945200 0.0002898550 0.0002852682 0.0002807580
#>          F2048        F2049        F2050        F2051        F2052        F2053
#> 1 0.0001184114 0.0001157214 0.0001131267 0.0001106236 0.0001082087 0.0001058784
#> 2 0.0001420425 0.0001390496 0.0001361509 0.0001333433 0.0001306236 0.0001279887
#> 3 0.0001693297 0.0001660188 0.0001627994 0.0001596686 0.0001566238 0.0001536622
#> 4 0.0002006033 0.0001969593 0.0001934022 0.0001899295 0.0001865391 0.0001832288
#> 5 0.0002361744 0.0002321813 0.0002282687 0.0002244347 0.0002206777 0.0002169960
#> 6 0.0002763233 0.0002719627 0.0002676748 0.0002634585 0.0002593124 0.0002552354
#>          F2054        F2055        F2056        F2057        F2058        F2059
#> 1 0.0001036298 0.0001014595 9.936475e-05 9.734269e-05 9.539062e-05 9.350598e-05
#> 2 0.0001254357 0.0001229618 1.205644e-04 1.182409e-04 1.159888e-04 1.138056e-04
#> 3 0.0001507813 0.0001479788 1.452522e-04 1.425993e-04 1.400178e-04 1.375055e-04
#> 4 0.0001799964 0.0001768398 1.737572e-04 1.707465e-04 1.678058e-04 1.649334e-04
#> 5 0.0002133878 0.0002098515 2.063856e-04 2.029884e-04 1.996585e-04 1.963944e-04
#> 6 0.0002512262 0.0002472836 2.434065e-04 2.395938e-04 2.358443e-04 2.321569e-04
#>          F2060        M2007        M2008        M2009        M2010        M2011
#> 1 0.0000916863 0.0009038192 0.0008846367 0.0008655757 0.0008466459 0.0008278568
#> 2 0.0001116892 0.0009440678 0.0009248942 0.0009058949 0.0008870752 0.0008684401
#> 3 0.0001350606 0.0009917700 0.0009724810 0.0009534185 0.0009345838 0.0009159782
#> 4 0.0001621275 0.0010478634 0.0010283288 0.0010090719 0.0009900908 0.0009713836
#> 5 0.0001931947 0.0011134843 0.0010935643 0.0010739725 0.0010547040 0.0010357539
#> 6 0.0002285305 0.0011900053 0.0011695476 0.0011494669 0.0011297558 0.0011104070
#>          M2012        M2013        M2014        M2015        M2016        M2017
#> 1 0.0008092175 0.0007907369 0.0007724234 0.0007542850 0.0007363294 0.0007185641
#> 2 0.0008499947 0.0008317434 0.0008136907 0.0007958408 0.0007781976 0.0007607649
#> 3 0.0008976030 0.0008794591 0.0008615475 0.0008438691 0.0008264244 0.0008092141
#> 4 0.0009529483 0.0009347828 0.0009168850 0.0008992529 0.0008818844 0.0008647772
#> 5 0.0010171175 0.0009987902 0.0009807673 0.0009630444 0.0009456168 0.0009284803
#> 6 0.0010914136 0.0010727687 0.0010544655 0.0010364973 0.0010188577 0.0010015404
#>          M2018        M2019        M2020        M2021        M2022        M2023
#> 1 0.0007009959 0.0006836316 0.0006664774 0.0006495392 0.0006328226 0.0006163327
#> 2 0.0007435460 0.0007265443 0.0007097627 0.0006932040 0.0006768709 0.0006607657
#> 3 0.0007922385 0.0007754981 0.0007589929 0.0007427231 0.0007266887 0.0007108895
#> 4 0.0008479294 0.0008313385 0.0008150025 0.0007989191 0.0007830860 0.0007675010
#> 5 0.0009116303 0.0008950626 0.0008787728 0.0008627569 0.0008470106 0.0008315299
#> 6 0.0009845390 0.0009678475 0.0009514599 0.0009353704 0.0009195732 0.0009040626
#>          M2024        M2025        M2026        M2027        M2028        M2029
#> 1 0.0006000744 0.0005840521 0.0005682701 0.0005527320 0.0005374413 0.0005224011
#> 2 0.0006448904 0.0006292471 0.0006138374 0.0005986629 0.0005837248 0.0005690243
#> 3 0.0006953253 0.0006799957 0.0006649004 0.0006500387 0.0006354100 0.0006210137
#> 4 0.0007521618 0.0007370660 0.0007222114 0.0007075956 0.0006932161 0.0006790708
#> 5 0.0008163106 0.0008013489 0.0007866407 0.0007721822 0.0007579695 0.0007439989
#> 6 0.0008888333 0.0008738797 0.0008591967 0.0008447789 0.0008306213 0.0008167190
#>          M2030        M2031        M2032        M2033        M2034        M2035
#> 1 0.0005076143 0.0004930831 0.0004788099 0.0004647964 0.0004510775 0.0004378458
#> 2 0.0005545622 0.0005403394 0.0005263563 0.0005126132 0.0004991373 0.0004860834
#> 3 0.0006068488 0.0005929145 0.0005792097 0.0005657335 0.0005525044 0.0005396365
#> 4 0.0006651572 0.0006514728 0.0006380154 0.0006247824 0.0006117841 0.0005990908
#> 5 0.0007302666 0.0007167690 0.0007035023 0.0006904631 0.0006776528 0.0006650965
#> 6 0.0008030671 0.0007896608 0.0007764954 0.0007635663 0.0007508663 0.0007383754
#>          M2036        M2037        M2038        M2039        M2040        M2041
#> 1 0.0004251136 0.0004128597 0.0004010641 0.0003897075 0.0003787718 0.0003682394
#> 2 0.0004734623 0.0004612579 0.0004494550 0.0004380387 0.0004269948 0.0004163098
#> 3 0.0005271377 0.0005149964 0.0005032011 0.0004917408 0.0004806049 0.0004697832
#> 4 0.0005867067 0.0005746236 0.0005628333 0.0005513280 0.0005401002 0.0005291424
#> 5 0.0006527939 0.0006407393 0.0006289275 0.0006173533 0.0006060115 0.0005948972
#> 6 0.0007260875 0.0007139995 0.0007021081 0.0006904103 0.0006789030 0.0006675831
#>          M2042        M2043        M2044        M2045        M2046        M2047
#> 1 0.0003580936 0.0003483186 0.0003388993 0.0003298210 0.0003210701 0.0003126332
#> 2 0.0004059706 0.0003959646 0.0003862798 0.0003769048 0.0003678283 0.0003590398
#> 3 0.0004592657 0.0004490429 0.0004391055 0.0004294448 0.0004200520 0.0004109188
#> 4 0.0005184475 0.0005080085 0.0004978186 0.0004878713 0.0004781601 0.0004686789
#> 5 0.0005840054 0.0005733314 0.0005628706 0.0005526183 0.0005425702 0.0005327219
#> 6 0.0006564476 0.0006454937 0.0006347185 0.0006241190 0.0006136926 0.0006034364
#>          M2048        M2049        M2050        M2051        M2052        M2053
#> 1 0.0003044978 0.0002966519 0.0002890838 0.0002817827 0.0002747379 0.0002679394
#> 2 0.0003505290 0.0003422860 0.0003343015 0.0003265663 0.0003190718 0.0003118094
#> 3 0.0004020373 0.0003933997 0.0003849984 0.0003768263 0.0003688763 0.0003611416
#> 4 0.0004594217 0.0004503825 0.0004415557 0.0004329358 0.0004245174 0.0004162953
#> 5 0.0005230692 0.0005136078 0.0005043338 0.0004952431 0.0004863319 0.0004775963
#> 6 0.0005933477 0.0005834240 0.0005736624 0.0005640606 0.0005546158 0.0005453256
#>          M2054        M2055        M2056        M2057        M2058        M2059
#> 1 0.0002613777 0.0002550435 0.0002489280 0.0002430228 0.0002373200 0.0002318116
#> 2 0.0003047713 0.0002979496 0.0002913369 0.0002849260 0.0002787100 0.0002726823
#> 3 0.0003536159 0.0003462926 0.0003391658 0.0003322295 0.0003254780 0.0003189057
#> 4 0.0004082643 0.0004004197 0.0003927565 0.0003852701 0.0003779559 0.0003708097
#> 5 0.0004690327 0.0004606375 0.0004524070 0.0004443379 0.0004364267 0.0004286701
#> 6 0.0005361876 0.0005271992 0.0005183582 0.0005096621 0.0005011087 0.0004926956
#>          M2060
#> 1 0.0002264905
#> 2 0.0002668366
#> 3 0.0003125075
#> 4 0.0003638270
#> 5 0.0004210649
#> 6 0.0004844206
head(frefictivetable)
#>       Id Gender DateOfBirth     DateIn    DateOut   Status
#> 1 100001 Female  1973-10-10 1996-01-01 2003-12-01    other
#> 2 100002   Male  1901-05-12 1996-01-01 2001-04-21 deceased
#> 3 100003 Female  1970-07-10 1996-01-01 2000-02-01    other
#> 4 100004   Male  1916-07-07 1996-01-01 2002-01-28 deceased
#> 5 100005 Female  1950-10-31 1996-01-01 2003-11-01    other
#> 6 100006   Male  1918-04-06 1996-01-01 2002-06-01 deceased

# (2) load of data
#

data(frefictivetable2)
range(frefictivetable2$DateIn)
#> [1] "1988-12-29" "1993-12-31"
range(frefictivetable2$DateOut)
#> [1] "1989-01-28" "1998-10-18"

# (3) other
#

if (FALSE) { # \dontrun{
data(freptfpermdis)
data(freptftempdis)

head(freptfpermdis)
head(freptftempdis)
} # }