Switzerland general liability triangles

swtri1auto is a named list of two triangles : the incurred (cumulative) amounts and the paid (cumulative) amounts.

Usage

data(swtri1auto)

Format

swtriangles is a named list of two matrices, respectively for incurred and paid amounts.

References

Dahms, R. (2008), A Loss Reserving Method for Incomplete Claim Data, Bulletin of the Swiss Association of Actuaries, pp. 127-148.

Dahms, R., Merz, M., Wuethrich, M.V. (2009), Claims development result for combined claims incurred and claims paid data. Bulletin Francais d'Actuariat 9 (18), 5-39.

Merz, M., and M. V. Wuethrich (2010), Paid-Incurred Chain Claims Reserving Method, Insurance: Mathematics and Economics 46, 2010, pp. 568-579, doi:10.1016/j.insmatheco.2010.02.004 .

Merz, M., and M. V. Wuethrich (2013), Estimation of Tail Development Factors in the Paid-Incurred Chain Reserving Method, Variance 71, pp. 61-73.

Examples

# (1) load of data
#
data(swtri1auto)