Normalized Hurricane Damages

Normalized Hurricane Damages in the United States: 1900-2005 used in Pielke et al. (2008). Originally, the data are stored in an Excel file with 4 worksheets. Damages are normalized according two approaches : (1) the methodology used by Pielke and Landsea (1998), adjusting for inflation, wealth, and population updated to 2005, called PL05; and (2) the methodology used by Collins and Lowe (2001), adjusting for inflation, wealth, and housing units updated to 2005, called CL05.

Usage

data(ushustormloss)
data(ushuannualloss)
data(ushuinflation)
data(ushupopulation)

Format

ushustormloss is a data frame of 7 columns and 207 rows:

Year

Year of the Hurricane.

Hurricane.Description

Description of the Hurricane.

State

States damaged by the Hurricane.

Category

Category of the Hurricane.

Base.Economic.Damage

Economic damages (original USD).

Normalized.PL05

Normalized PL05 damages (2005 USD).

Normalized.CL05

Normalized CL05 damages (2005 USD).

ushuannualloss is a data frame of 2 columns and 106 rows:

Year

Year.

Normalized.PL05

Total year Normalized damages (2005 USD).

ushuinflation is a data frame of 9 columns and 106 rows:

Year

Year.

Implicit.Price.Deflator

Implicit price deflator.

Inflation.Multiplier

Inflation multiplier.

Wealth

Wealth.

Real.Wealth.2005.Base

Real wealth (2005 base).

Real.Wealth.Per.Capita

Real wealth per capita.

Real.Wealth.Per.Capita.Multiplier

Real wealth per capita multiplier.

Real.Wealth.Per.Housing.Unit

Real wealth per housing unit.

Real.Wealth.Per.Housing.Unit.Multiplier

Real wealth per housing multiplier.

ushupopulation is a data frame of 12 columns and 217 rows:

Storm.ID

Storm ID.

Storm.Year

Year of the Storm.

Storm.Name

Name of the Storm.

County.Original.Population

Original population in counties affected by storm.

County.2005.Population

2005 population in counties affected by storm.

County.Population.Multiplier

County population multiplier.

County.Original.Housing.Units

Original housing units in counties affected by storm.

County.2005.Housing.Units

2005 housing units in counties affected by storm.

Housing.Units.Multiplier

Housing units multiplier.

Year

Year

US.Population

Total US population.

US.Housing.Units

Total US housing units.

Source

https://sciencepolicy.colorado.edu/publications/special/normalized_hurricane_damages.html

References

Pielke, Gratz, Landsea, Collins, Saunders, and Musulin (2008), Normalized Hurricane Damages in the United States: 1900-2005, Natural Hazards Review, Volume 9, Issue 1, pp. 29-42. doi:10.1061/(ASCE)1527-6988(2008)9:1(29)

Examples

# (1) load of data
#
data(ushustormloss)